Don't attempt to include pkcs#11 support if cgo is disabled
It is disabled by default when cross compiling.
