+++ .gitignore
@@ -1,2 +1,2 @@
 build/
-gokeyless
+/gokeyless
