+++ cmd/gokeyless/.gitignore
@@ -1,0 +1,2 @@
+/gokeyless
+/gokeyless.yaml
