+++ cmd/gokeyless/cgo.go
@@ -1,0 +1,13 @@
+// +build cgo
+
+package main
+
+import (
+	"crypto"
+
+	"github.com/cloudflare/gokeyless/server"
+)
+
+func loadURI(uri string) (crypto.Signer, error) {
+	return server.DefaultLoadURI(uri)
+}
