+++ internal/rfc7512/rfc7512.go
@@ -1,3 +1,5 @@
+// +build cgo
+
 // Package rfc7512 provides a parser for the PKCS #11 URI format as specified in
 // RFC 7512: The PKCS #11 URI Scheme. Additionally, it provides a wrapper around
 // the crypto11 package for loading a key pair as a crypto.Signer object.
